requirements.txt
alembic==1.13.1
arrow==1.3.0
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
infinity==1.5
intervals==0.9.2
itsdangerous==2.2.0
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
packaging==24.0
pip==24.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
setuptools==69.5.1
six==1.16.0
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
validators==0.28.1
Werkzeug==3.0.2
wheel==0.43.0
WTForms==3.1.2
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
